compute_scripts/requirements.txt
antlr4_python3_runtime==4.9.3+computecanada
distlib @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/distlib/distlib-0.3.7
editables @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/editables/editables-0.5
filelock @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/filelock/filelock-3.12.2
flit_core @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/flit_core/flit_core-3.9.0
fsspec==2024.10.0+computecanada
hatch-vcs @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/hatch_vcs/hatch_vcs-0.3.0
hatchling @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/hatchling/hatchling-1.18.0
hydra_core==1.3.2+computecanada
iopath==0.1.10
jinja2==3.1.4+computecanada
MarkupSafe==2.1.5+computecanada
-e git+https://github.com/kew6688/MicroNuclei_Detection.git@347b4514526a7c18caf681678434ea2e20d626cb#egg=mn_segmentation
mpmath==1.3.0+computecanada
networkx==3.4.2+computecanada
numpy==2.1.1+computecanada
omegaconf==2.3.0+computecanada
packaging @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/packaging/packaging-23.1
pathspec @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/pathspec/pathspec-0.11.2
pillow==10.4.0+computecanada
Pillow_SIMD==9.5.0.post2+computecanada
platformdirs @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/platformdirs/platformdirs-3.10.0
pluggy @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/pluggy/pluggy-1.3.0
portalocker==3.0.0
pyparsing @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/pyparsing/pyparsing-3.1.1
PyYAML==6.0.2+computecanada
-e git+https://github.com/facebookresearch/sam2.git@722d1d15111c689908aeeb82d49a57780aac5153#egg=SAM_2
setuptools-scm @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/setuptools_scm/setuptools_scm-7.1.0
six @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/six/six-1.16.0
sympy==1.13.1+computecanada
tomli @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/tomli/tomli-2.0.1
torch==2.5.1+computecanada
torchvision==0.20.1+computecanada
tqdm==4.67.1+computecanada
trove-classifiers @ file:///cvmfs/soft.computecanada.ca/easybuild/sources/p/Python/extensions/trove_classifiers-2023.8.7-py3-none-any.whl#sha256=a676626a31286130d56de2ea1232484df97c567eb429d56cfcb0637e681ecf09
typing_extensions==4.12.2+computecanada
virtualenv @ file:///tmp/ebuser/avx2/Python/3.11.5/GCCcore-12.3.0/virtualenv/virtualenv-20.24.3
